Rename explore/VenueCard -> explore/ExploreCard, add TODO comments, move VenueReactButton to shared component level
