+++ src/components/explore/ExploreCard.styles.ts
